Module Kicadsch.MakeSchPainter

A module able to paint a schematic file in a painter context

Parameters

module P : Sigs.Painter

Signature

type schContext

the schematic context

val initial_context : ?allow_missing_component:bool -> Sigs.revision -> schContext

initial_context allow_missing_component revision

returns
an new empty context

val add_lib : string -> schContext -> schContext

add_lib line context parse the content of line provided to libs to the context.

returns
the updated context

val parse_line : Stdlib.String.t -> schContext -> schContext

parse_line line context parse a new line of schematic and update context.

returns
the updated context

val output_context : schContext -> P.t

output_context context output write the context as a image format to output