Module Kicadsch.Sigs

Kicad modules Signatures

type orientation =

| Orient_H

| Orient_V

orientation of a text

type coord =

| Coord of int * int

absolute coordinates in the drawing

type size =

| Size of int

font size of a text

type justify =

| J_left

| J_right

| J_center

| J_bottom

| J_top

Text justification of a text

type style =

| Bold

| Italic

| BoldItalic

| NoStyle

Style of a text

type kolor = [

| `NoColor

| `Black

| `Green

| `Red

| `Blue

| `Brown

]

Color of the text. These are the colors appearing in Kicad schematics

type transfo = (int * int) * (int * int)

Transformation matrix of a relative coordinate around an absolute coordinate. The matrix is layed out as a pair of lines of pairs

type revision =

| First of string

| Second of string

| No_Rev

module type Painter = sig ... end

A module able to paint a canvas with several graphic primitives and then to process the canvas into a picture file format. The functions are supposed to be pure

module type SchPainter = sig ... end

A module able to paint a schematic file in a painter context

module type CompPainter = sig ... end

The library that is able to read component libraries and memorize the read components. Then when passed a drawing context and a component to paint it can paint the component on demand to the drawing context