Module type Sigs.CompPainter

The library that is able to read component libraries and memorize the read components. Then when passed a drawing context and a component to paint it can paint the component on demand to the drawing context

type t

A component Library manager

type drawContext

A drawing context

val lib : unit -> t

lib ()

returns
an empty new component manager

val append_lib : string -> t -> t

append_lib stream context appends the lib contained in the stream to the context.

returns
the updated context

val plot_comp : t -> string -> int -> coord -> transfo -> bool -> drawContext -> drawContext * bool

plot_comp lib name partnumber origin transformation allow_missing context find in lib the component with given name and plot the part partnumber at origin after transfomation into the graphical context and the fact that the component is multipart. If the component is not found, raise an exception, unless allow_missing is true.

returns
the updated graphical context