Module Kicadsch__.Schparse

val create_lib_parse_fun : name:string -> regexp_str:('a, Stdlib.Scanf.Scanning.in_channel, 'b, 'c -> 'd option, 'a -> 'e, 'e) Stdlib.format6 -> processing:'c -> string -> 'd option

This function generates a parsing function which outputs an 'a option Note that some lines may not yield any correct output, so the output is an option. *

val create_parse_fun : name:string -> regexp_str:('a, Stdlib.Scanf.Scanning.in_channel, 'b, 'c -> 'd option, 'a -> 'e, 'e) Stdlib.format6 -> extract_fun:'c -> string -> onerror:(unit -> 'f) -> process:('d -> 'f) -> 'f

val parse_list : ?⁠cond:('a -> bool) -> ('b, Stdlib.Scanf.Scanning.in_channel, 'c, ('d -> 'd) -> 'a, 'b -> 'e, 'e) Stdlib.format6 -> string -> 'a list